Agrego chekeo login de usuario con db

+++ GTBusinessLayer/SalesDetail.cs
@@ -7,7 +7,7 @@ namespace GTBusinessLayer
     public class SalesDetail
     {
         public Product _product { get; private set; }
-        public double TotalPrice { get; private set; }
+        public decimal TotalPrice { get; private set; }
         public int TotalItems { get; private set; }
